\name{GrowingSOM}
\alias{GSOM}
\alias{GrowingSelfOrganizingMap}
\docType{package}
\title{
\packageTitle{GrowingSOM}
}
\description{
\packageDescription{GrowingSOM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GrowingSOM}
\packageIndices{GrowingSOM}
%%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GrowingSOM}

Maintainer: \packageMaintainer{GrowingSOM}
}
\references{
%%~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
